+++ src/siganalysis/__init__.py
@@ -4,6 +4,7 @@
 # can be found in the LICENSE.txt file for the project.
 from .siganalysis import (
     STFT_WINDOWS,
+    WINDOW_FUNCTIONS,
     calculate_peak_hold,
     freq_bin,
     hz2khz,
@@ -18,6 +19,7 @@ from .siganalysis import (
 
 __all__ = [
     "STFT_WINDOWS",
+    "WINDOW_FUNCTIONS",
     "calculate_peak_hold",
     "freq_bin",
     "hz2khz",